Fix inherited userinfo leaking when parsing a new authority in Url
ParseAuthority and ParseFileHost now clear _username/_password at entry
so base URL credentials don't persist when the input specifies a fresh
authority without userinfo.
