+++ src/AngleSharp/Dom/Url.cs
@@ -124,7 +124,7 @@ namespace AngleSharp.Dom
             _host = address._host;
             _username = address._username;
             _password = address._password;
-             _relative = address._relative;
+            _relative = address._relative;
             _hasPath = address._hasPath;
             _schemeData = address._schemeData;
         }
@@ -1650,6 +1650,6 @@ namespace AngleSharp.Dom
                 return chars.Slice(0, count).ToString();
             }
         }
-#endregion
+        #endregion
     }
 }
